third_party/WebKit/LayoutTests/http/tests/sendbeacon/beacon-allowance.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  let result = true;
  for (var i = 0 ; result && i < 100; i++)
    result = navigator.sendBeacon("resources/blank.txt", new Uint32Array(1000));
  assert_false(result);
}, "Test navigator.sendBeacon() restrictions on (accumulated) transmission length.");
</script>
